requirements.txt
numpy
pandas
Pillow
opencv-python
trimesh
open3d
networkx
scipy
fast_slic
numba
pycuda
